requirements.txt
asgiref==3.8.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
python-decouple==3.8
sqlparse==0.4.4
